+++ docs/changes.rst
@@ -5,6 +5,9 @@ Weblate 2026.6
 
 .. rubric:: New features
 
+* :doc:`/admin/announcements` can now also be managed via the :ref:`api` for specific project languages.
+* Team memberships can now be limited to selected languages for per-user translation permissions.
+
 .. rubric:: Improvements
 
 * Docker containers can now adjust :setting:`WEBLATE_FORMATS`.